2025-12-22first commit

+++ service/service-dispatch/src/main/java/com/atguigu/daijia/dispatch/xxl/job/JobHandler.java
@@ -38,8 +38,7 @@ public class JobHandler {
             e.printStackTrace();
         } finally {
             long times = System.currentTimeMillis()- startTime;
-            //TODO 完善long
-            xxlJobLog.setTimes((int)times);
+            xxlJobLog.setTimes(times);
             xxlJobLogMapper.insert(xxlJobLog);
         }
     }
